+++ src/ralph/models/xapi/concepts/verbs/adl_vocabulary.py
@@ -37,3 +37,17 @@ class AnsweredVerb(BaseXapiVerb):
         "http://adlnet.gov/expapi/verbs/answered"
     ] = "http://adlnet.gov/expapi/verbs/answered"
     display: Optional[Dict[Literal[LANG_EN_US_DISPLAY], Literal["answered"]]]
+
+
+class RegisteredVerb(BaseXapiVerb):
+    """Pydantic model for registered `verb`.
+
+    Attributes:
+        id (str): Consists of the value `http://adlnet.gov/expapi/verbs/registered`.
+        display (dict): Consists of the dictionary `{"en-US": "registered"}`.
+    """
+
+    id: Literal[
+        "http://adlnet.gov/expapi/verbs/registered"
+    ] = "http://adlnet.gov/expapi/verbs/registered"
+    display: Optional[Dict[Literal[LANG_EN_US_DISPLAY], Literal["registered"]]]